Prevent mdm_worker_class method from unnecessarily creating classes
The previous implementation would create a new class every time the
method was called. It also generated a warning when it was called more
than once, because it was reassigning an already assigned constant.

Under normal circumstances, this method would probably never be called
more than once, but when a Rails application reloads, and a model that
defines a global_registry_binding is reloaded, that's exactly what
happens. This causes warnings to display in the console (or test output)
any time the application has been reloaded (which, if you're using
Spring, is every time you run a command).

In addition to short circuting by returning the originally defined
constant, I also added a call to

    ActiveSupport::Dependencies.mark_for_unload

Which tells Rails to undefine this constant when the app is being
reloaded, ensuring that if the app is being reloaded due to
configuration changes to the global_registry_binding, the changes will
correctly be picked up.